Removed some duplication
